Add new interface model for Axe.Windows.Automation
The model has not been implemented yet, except for the Config class and the ScannerFactory class.

Changes since the last design review:
- Change ScanResult.OutputFiles from an IEnumerable<string> to a Tuple: (string A11yTest, string Sarif) so that output files are grouped together, but still distinguishable by type
- Removed Scanner.Builder and created ScannerFactory since creating new objects which implement a specific interface is the job of a factory. It also hides the implementation of Scanner entirely.
- For the returned data classes, changed private access to internal access because we don't need to manage construction of these classes with a builder.
